# docs/openapi/be-plan-phase4.openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "AgriDrone.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/auth/invitations/preview": {
      "post": {
        "tags": [
          "TenantInvitation"
        ],
        "summary": "Xem trước lời mời và xác định có cần tạo tài khoản.",
        "description": "Token phải còn hiệu lực và thuộc tenant đang hoạt động. Endpoint chỉ trả\r\nemail đã che một phần để frontend chọn đúng biểu mẫu accept invitation.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewTenantInvitationRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewTenantInvitationRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewTenantInvitationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/invitations/accept": {
      "post": {
        "tags": [
          "TenantInvitation"
        ],
        "summary": "Chấp nhận lời mời tham gia tenant.",
        "description": "Xác thực invitation token, tạo tài khoản nếu cần và kích hoạt tenant\r\nmembership theo vai trò đã được mời. Token chỉ được sử dụng một lần.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptTenantInvitationRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptTenantInvitationRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptTenantInvitationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/users": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Lấy danh sách người dùng toàn hệ thống.",
        "description": "Trả danh sách phân trang dành cho System Admin, không giới hạn theo\r\ntenant đang chọn.",
        "parameters": [
          {
            "name": "PageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/current/profile": {
      "put": {
        "tags": [
          "User"
        ],
        "summary": "Cập nhật hồ sơ người dùng hiện tại.",
        "description": "Cập nhật họ tên và số điện thoại của người dùng đang đăng nhập; email\r\nvà quyền truy cập không thay đổi qua endpoint này.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserProfileRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserProfileRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserProfileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/current/change-password": {
      "put": {
        "tags": [
          "User"
        ],
        "summary": "Đổi mật khẩu của người dùng hiện tại.",
        "description": "Xác minh mật khẩu cũ trước khi lưu mật khẩu mới cho tài khoản đang\r\nđăng nhập.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserPasswordRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserPasswordRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserPasswordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AcceptTenantInvitationRequest": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PreviewTenantInvitationRequest": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpdateUserPasswordRequest": {
        "type": "object",
        "properties": {
          "oldPassword": {
            "type": "string",
            "nullable": true
          },
          "newPassword": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UpdateUserProfileRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Enter JWT token",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}